backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
sendgrid==6.11.0
requests==2.31.0
gunicorn==21.2.0
web3==6.11.3
eth-account==0.10.0
cryptography==41.0.7
python-gnupg==0.5.1
Flask-Limiter==3.5.0
SQLAlchemy==2.0.23
pyotp==2.9.0
qrcode==7.4.2
pywebpush==1.14.1

